fix(objcinteropge): default values for nullable types
Default values of fields and variables of pointer-to-class types were printed
(in the experimental and generate-definitions modes) always as `Option<...>.None.getOrThrow()`.

That is OK as long this code is not intended to run.

But this can be a problem if using in the normal mode.

Partially fixed, for the case of nullable object pointers (which are converted to `Option`).

Their default value is correctly `None` now.

Signed-off-by: Mikhail Nesterov <nesterov.mikhail2@huawei-partners.com>